Add LinkedContentPage page type
