+++ content/templates/content/linked_content_page.html
@@ -1,0 +1,1 @@
+{% extends "content/content_page.html" %}
